requirements.txt
websockets
avro-python3
schedule
confluent_kafka
urllib3
requests